+++ exts/isaacsim.zmq.bridge.examples/docs/CHANGELOG.md
@@ -3,5 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [1.2.0] - 2026-05-19
+- Migrate from deprecated `isaacsim.core.api` to experimental APIs for Isaac Sim 6.0
+
+## [1.1.0] - 2025-03-03
+- Added headless standalone mode support
+
 ## [1.0.0] - 2025-03-03
 - Initial version